+++ src/gameObject/GameObject.hpp
@@ -21,16 +21,16 @@ class GameObject {
     void addSprite(string fileName);
 
     int addAnimation();
-    void addAnimationFrame(long index, long x, long y, long width, long height);
-    void addAnimationFrames(long index, long fromX, long fromY, long toX, long toY, long width, long height);
+    void addAnimationFrame(int index, int x, int y, int width, int height);
+    void addAnimationFrames(int index, int fromX, int fromY, int toX, int toY, int width, int height);
     Animation* getAnimation(int index);
 
     void draw(sf::RenderWindow* window);
 
-    long getActiveAnimationLocalIndex();
-    long getActiveAnimationGlobalIndex();
+    int getActiveAnimationLocalIndex();
+    int getActiveAnimationGlobalIndex();
 
-    void playAnimation(long globalIndex, long localIndex);
+    void playAnimation(int globalIndex, int localIndex);
     void stopAnimation();
     void pauseAnimation();
     void resumeAnimation();
@@ -38,8 +38,8 @@ class GameObject {
 
     private:
     int globalIndex;
-    long activeAnimationLocalIndex;
-    long activeAnimationGlobalIndex;
+    int activeAnimationLocalIndex;
+    int activeAnimationGlobalIndex;
     sf::Vector2f position;
     sf::Sprite sprite;
     sf::Texture texture;
